requirements.txt
streamlit
pillow
tensorflow
numpy
matplotlib
